DEU eursf
@michihdeu - these were mostly fine, I've tweaked the names, added
concurrent sections with deub and 'super-2' sections (where routes make
it seem silly not to), but feel free to change that back if you want.
City names appended onto streets need to be in the Abbrev field, rather
than the route one, and I feel that AST is a banner of sorts so should
really go there.