usaca CA 260 truncation
This is part of an oddball situation in Oakland I'll discuss on the
forum. It's clear enough CA 260 should be truncated if it remains in the
HB, though what to do with concurrent route CA 61 (including the mileage
removed from CA 260, and perhaps absorbing the rest of CA 260) is less
clear.