PA 949: Recentering
Replaced BowRd woth Dickey Rd (DicRd).  Replaced SweRd and EmpRidRd with Arroyo Rd (AroRd).  As ArkLn is a gated entrance, and not even a named road in the SLDs, I made it a hidden shaping point making ArkLn an alternate-label for CarRd.  This adds 5 miles to a the mileage of a couple users though.