Creating PA 423 TRUCK (Pocono Pines)
Found while clinching PA 940 on way to Morris County, NJ, Roadmeet.  GSV shows it posted on PA 940 EB to I-380 NB.  Both GSV and the VideoLog do not show it on PA 423 SB at I-380 at this time, but I believe that I know enough to see that there is no way that it would go past I-380 (Exit 8).