BW E54
When I drove B31 in 2016, I saw E54 signed and changed the route in HB.
However, I drove B31n multiple times this year and E54 is clearly signed
on B31n (opened in 1980s). It is the direct route with much higher road
standard etc. and I think this is the one where E54 should be.