BW Albstadt
The situation was always odd and it was difficult to check km posts
on-site (well, I only checked the "northern" streets in 2020 (or early
2021) and there was nothing within the town boundaries) but but OSM +
geoportal are updated and both indicate what "makes sense". I might to
check both streets parallel to B463 but hey, it already looks better
than before! I'm still the only traveler on all those roads...