CZE D7 Louny opened
https://www.skyscrapercity.com/threads/1464920/post-185367996

Keeping I7 on concurrency since it is just an upgrade on same alignment and the next section to the east will be opened by end of the year, i.e. it makes no sense to split the I7 route and create a new one for just 3 months including breaking list files.