TUR Istanbul (tursf)
Sources for route name and exit numbering of HabHasBag (Cebeci tunnel)

https://www.trthaber.com/haber/ekonomi/kuzey-marmara-otoyolunun-son-kesimi-acilisa-hazir-576111.html
https://youtu.be/mdpGfLln4a8
There are also O7 km posts and "To O7" signs westbound. I failed to find a video which was recorded "eastbound" (but many westbound) so that I could check if O7 is also indicated eastbound and thus the route would belong to turo system. Nonetheless, OSM does not indicate O7 - but GM :-1:  GSV is only partially available but was recorded before the opening. The exit numbering has obviously changed afterwards.